Move phases in preparation for refactor
