+++ src/game/game_state.cpp
@@ -1,10 +1,8 @@
 #include "game_state.h"
-#include "events.h"
 
-MenuPhase GameState::menu_phase;
 BuildPhase GameState::build_phase;
 DungeonPhase GameState::dungeon_phase;
-MinigamePhase GameState::minigame_phase;
+
 Phase* GameState::curr_phase;
 std::map<int, Player*> GameState::players;
 
@@ -67,7 +65,7 @@ void GameState::set_phase(Phase* phase) {
 void GameState::set_phase(proto::Phase phase) {
     switch (phase) {
     case proto::Phase::MENU:
-        GameState::set_phase(&GameState::menu_phase);
+        // FIXME(metakirby5)
         break;
     case proto::Phase::BUILD:
         GameState::set_phase(&GameState::build_phase);
@@ -76,7 +74,7 @@ void GameState::set_phase(proto::Phase phase) {
         GameState::set_phase(&GameState::dungeon_phase);
         break;
     case proto::Phase::MINIGAME:
-        GameState::set_phase(&GameState::minigame_phase);
+        // FIXME(metakirby5)
         break;
     }
 }
