+++ src/game/phase/dungeon.cpp
@@ -1,0 +1,22 @@
+#include "dungeon.h"
+
+void DungeonPhase::setup() {
+    joystick_conn = events::stick_event.connect([&](events::StickData d) {
+        // Left stick
+        if (d.input == events::STICK_LEFT) {
+            events::dungeon::network_player_move_event(d);
+        }
+    });
+
+    button_conn = events::button_event.connect([&](events::ButtonData d) {
+        // A button pressed.
+        if (d.input == events::BTN_A && d.state == 1) {
+            events::dungeon::player_interact_event();
+        }
+    });
+}
+
+void DungeonPhase::teardown() {
+    joystick_conn.disconnect();
+    button_conn.disconnect();
+}
