+++ src/game/phase/dungeon.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "phase.h"
+
+class DungeonPhase : public TimedPhase {
+public:
+    void setup();
+    void teardown();
+private:
+    boost::signals2::connection joystick_conn;
+    boost::signals2::connection button_conn;
+};
